MongoDB + разнес по модулям код
