+++ index.js
@@ -1,7 +1,8 @@
 const express = require('express');
 const app = express();
 const port = 8000;
-
+const db = require('../strada_roadmap/db')
+const service = require('../strada_roadmap/service')
 app.use(express.json())
 
 app.get('/', (req, res) => {
@@ -9,19 +10,25 @@ app.get('/', (req, res) => {
 })
 
 app.get('/tasks', (req, res) => {
-  res.send(showList());
+  // res.send(showList());
+  res.send(service.showList())
 })
 
-app.post('/tasks', (req, res) => {
-  const {name, status = 'to do', priority = 'low' } = req.body;
-  addTask(name, status, priority)
-  res.send(`Task added ${name}`);
+app.post('/tasks', async (req, res) => {
+  try {
+    const {name, status = 'to do', priority = 'low' } = req.body;
+    const task = {name, status, priority}
+    await db.getDB().collection('tasks').insertOne(task)
+    res.send(`Task added ${name}`);
+  } catch (e) {
+    console.error(e);
+  }
 })
 
 app.delete('/tasks', (req, res) => {
   const {name} = req.body;
   if (name) {
-    const isDeleted = deleteTask(name);
+    const isDeleted = service.deleteTask(name);
     if (isDeleted){
       res.send(`Task deleted ${name}`)
     } else 
@@ -34,7 +41,7 @@ app.delete('/tasks', (req, res) => {
 app.put('/tasks/', (req, res) => {
   const {name, status } = req.body;
   if (name) {
-    const isChanged = changeStatus(name, status);
+    const isChanged = service.changeStatus(name, status);
     if (isChanged) {
       res.send(`Task was changed ${name} for status: ${status}`);
     } else {
@@ -49,37 +56,4 @@ app.listen(port, () => {
   console.log(`Example app listening on port ${port}`);
 })
 
-const list = [
-  {name: 'first task', status: 'in progress', priority: 'low'},
-  {name: 'second task', status: 'to do', priority: 'high'},
-
-]
-
-const addTask = (nameTask, statusTask = 'to do', priorityTask = 'low') => {
-  list.push({name : nameTask, status : statusTask, priority : priorityTask});
-}
-
-const deleteTask = (nameTask) => {
-  const indexDeletedTask = list.findIndex(task => task.name === nameTask);
-  if (indexDeletedTask !== 1) {
-    list.splice(indexDeletedTask, 1);
-    return true
-  } else {
-    return false
-  }
-}
-
-const changeStatus = (nameTask, statusTask) => {
-  const findTask = list.find(task => task.name === nameTask);
-  if (findTask) {
-    findTask.status = statusTask;
-    return true
-  } else {
-    return false
-  }
-}
-
-const showList = () => {
-  return list;
-}
