+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "dependencies": {
     "date-fns": "^3.6.0",
-    "express": "^4.21.0"
+    "express": "^4.21.0",
+    "mongodb": "^6.9.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
